+++ public/icons/logo-types.png
