+++ src/app/help-center/page.tsx
@@ -1,4 +1,4 @@
-import HelpCenter from "../../components/HelpCenter";
+import HelpCenter from "../../components/help-center/help-center";
 
 export default function HelpCenterPage() {
   return <HelpCenter />;
