+++ src/components/help-center/help-center.tsx
@@ -5,7 +5,7 @@ import {
   Coins,
   ShieldCheck,
 } from "lucide-react";
-import { Input } from "@/components/ui/input";
+import { Input } from "@/components/common/input";
 
 export default function HelpCenter() {
   const sections = [
