+++ src/components/marketplace/filter-section.tsx
@@ -1,8 +1,8 @@
 "use client"
 
-import { Button } from "@/components/ui/button"
-import { Checkbox } from "@/components/ui/checkbox"
-import { Slider } from "@/components/ui/slider"
+import { Button } from "@/components/button"
+import { Checkbox } from "@/components/check-box"
+import { Slider } from "@/components/common/slider"
 
 interface FilterSectionProps {
   priceRange: number[];
