+++ src/components/marketplace/marketplace-card.tsx
@@ -1,8 +1,8 @@
 "use client"
 
 import { Star } from "lucide-react"
-import { Button } from "@/components/ui/button"
-import { Card, CardContent, CardFooter } from "@/components/ui/card"
+import { Button } from "@/components/button"
+import { Card, CardContent, CardFooter } from "@/components/cards/card"
 import { cn } from "@/lib/utils"
 import Image from "next/image"
