+++ src/components/marketplace/sorting-buttons.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { Button } from "@/components/ui/button"
+import { Button } from "@/components/button"
 
 interface SortingButtonsProps {
   onSort: (type: 'new' | 'priceAsc' | 'priceDesc' | 'rarity') => void;
