+++ src/components/misc/guns.tsx
