+++ src/components/ui/icons/braavos-icon.tsx
